pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-server-sqlite"
version = "0.1.0"
description = "Servidor MCP en Python con Clean Architecture y base de datos SQLite"
readme = "README.md"
license = {text = "MIT"}
authors = [
    { name = "User", email = "user@example.com" }
]
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "pydantic>=2.0.0",
    "motor>=3.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "autopep8>=2.0.0",
    "pytest>=7.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
mcp-server-sqlite = "main:mcp"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.autopep8]
max_line_length = 79
aggressive = 1

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: tests que toman mucho tiempo (ej. rate-limit)",
]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
